chore: move getFixedT to web (#13446)
